src/crawler/requirements.txt
beautifulsoup4==4.14.3
requests==2.32.5
psycopg==3.3.3
google-cloud-storage==2.19.0